Refactor to use shared models and enums
This commit refactors the codebase to utilize shared models and enums from the `AspireCafe.Shared` namespace, replacing previous domain-specific implementations in both the `CounterApi` and `ProductApi` domains.

Key changes include:
- Updated test files to standardize the testing framework.
- Added new dependencies for `eslint` in `package.json` and `package-lock.json` to enhance code quality.
- Replaced domain-specific service and view models in controllers and business logic with shared counterparts.
- Consolidated enumeration definitions into `CounterEnumerations.cs`.
- Updated interfaces (`IData`, `IBusiness`, `IFacade`) to align with the new shared models.
- Restructured `Order` and `Product` domain models for improved consistency.

These changes aim to streamline the codebase, improve maintainability, and enhance clarity in data structures.
