+++ src/Aspirecafe/AspireCafe.CounterApi.TestLayer/ControllerTests.cs
@@ -1,8 +1,8 @@
 ﻿using AspireCafe.CounterApi.Controllers;
 using AspireCafe.CounterApiDomainLayer.Facade;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.Service;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.View;
 using AspireCafe.Shared.Enums;
+using AspireCafe.Shared.Models.Service.Counter;
+using AspireCafe.Shared.Models.View.Counter;
 using AspireCafe.Shared.Results;
 using Moq;
