+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Business/IBusiness.cs
@@ -1,11 +1,5 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Service;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.View;
-using Microsoft.Azure.Cosmos.Serialization.HybridRow;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AspireCafe.Shared.Models.Service.Counter;
+using AspireCafe.Shared.Models.View.Counter;
 
 namespace AspireCafe.CounterApiDomainLayer.Business
 {
