+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Facade/Facade.cs
@@ -1,15 +1,9 @@
 ﻿using AspireCafe.CounterApiDomainLayer.Business;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.Service;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.View;
 using AspireCafe.CounterApiDomainLayer.Managers.Validators;
 using AspireCafe.Shared.Enums;
+using AspireCafe.Shared.Models.Service.Counter;
+using AspireCafe.Shared.Models.View.Counter;
 using AspireCafe.Shared.Results;
-using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.CounterApiDomainLayer.Facade
 {
