+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Managers/Extensions/CounterExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Domain;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.Enums;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.Service;
-using AspireCafe.CounterApiDomainLayer.Managers.Models.View;
+using AspireCafe.Shared.Models.Service.Counter;
+using AspireCafe.Shared.Models.View.Counter;
 
 namespace AspireCafe.CounterApiDomainLayer.Managers.Extensions
 {
