+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Managers/Models/Domain/OrderDomainModel.cs
@@ -1,9 +1,4 @@
 ﻿using AspireCafe.Shared.Models.Domain;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Domain
 {
