+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Managers/Models/Domain/OrderFooterDomainModel.cs
@@ -1,9 +1,4 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AspireCafe.Shared.Enums;
 
 namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Domain
 {
