+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Managers/Models/Domain/OrderLineItemDomainModel.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Domain
+﻿namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Domain
 {
     public class OrderLineItemDomainModel
     {
