+++ src/Aspirecafe/Aspirecafe.Counterapidomainlayer/Managers/Validators/OrderViewModelValidator.cs
@@ -1,10 +1,5 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.View;
+﻿using AspireCafe.Shared.Models.View.Counter;
 using FluentValidation;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.CounterApiDomainLayer.Managers.Validators
 {
