+++ src/Aspirecafe/Aspirecafe.Productapi/Controllers/CatlogController.cs
@@ -1,9 +1,9 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Facade;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
+using AspireCafe.Shared.Extensions;
+using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
 using AspireCafe.Shared.Results;
 using Microsoft.AspNetCore.Mvc;
-using AspireCafe.Shared.Extensions;
 
 namespace AspireCafe.ProductApi.Controllers
 {
