+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Business/Business.cs
@@ -1,7 +1,7 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Data;
 using AspireCafe.ProductApiDomainLayer.Managers.Extensions;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
+using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
 
 namespace AspireCafe.ProductApiDomainLayer.Business
 {
