+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Business/IBusiness.cs
@@ -1,6 +1,5 @@
-﻿using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
-using AspireCafe.Shared.Results;
+﻿using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
 
 namespace AspireCafe.ProductApiDomainLayer.Business
 {
