+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Data/CatalogData.cs
@@ -1,13 +1,7 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Managers.Context;
 using AspireCafe.ProductApiDomainLayer.Managers.Models.Domain;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
+using AspireCafe.Shared.Models.View.Product;
 using Microsoft.EntityFrameworkCore;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.ProductApiDomainLayer.Data
 {
