+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Data/IData.cs
@@ -1,11 +1,4 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Managers.Models.Domain;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.ProductApiDomainLayer.Data
 {
