+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Facade/CatalogFacade.cs
@@ -1,14 +1,9 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Business;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
 using AspireCafe.Shared.Cache;
+using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
 using AspireCafe.Shared.Results;
 using Microsoft.Extensions.Caching.Distributed;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace AspireCafe.ProductApiDomainLayer.Facade
 {
