+++ src/Aspirecafe/Aspirecafe.Productapidomainlayer/Managers/Extensions/ProductExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using AspireCafe.ProductApiDomainLayer.Managers.Models.Domain;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Enums;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.Service;
-using AspireCafe.ProductApiDomainLayer.Managers.Models.View;
+using AspireCafe.Shared.Enums;
+using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
 
 namespace AspireCafe.ProductApiDomainLayer.Managers.Extensions
 {
