+++ src/Aspirecafe/Aspirecafe.Shared/Enums/Enumerations.cs
@@ -24,5 +24,63 @@ namespace AspireCafe.Shared.Enums
         Forbidden,
         InternalServerError
     }
+
+    public enum ProductType
+    {
+        Food,
+        Beverage,
+        Dessert,
+        Snack
+    }
+
+    public enum ProductCategory
+    {
+        Appetizer,
+        MainCourse,
+        Dessert,
+        Beverage,
+        Snack
+    }
+
+    public enum ProductStatus
+    {
+        Available,
+        Unavailable,
+        OutOfStock
+    }
+
+    public enum RouteType
+    {
+        Barista,
+        Kitchen
+    }
+
+    public enum OrderType
+    {
+        DineIn = 1,
+        TakeAway = 2,
+        Delivery = 3
+    }
+
+    public enum PaymentMethod
+    {
+        Cash = 1,
+        Card = 2,
+        MobilePayment = 3
+    }
+
+    public enum PaymentStatus
+    {
+        Paid = 1,
+        Unpaid = 2,
+        Refunded = 3
+    }
+
+    public enum OrderStatus
+    {
+        Pending = 1,
+        Completed = 2,
+        Cancelled = 3
+    }
 }
