+++ src/Aspirecafe/Aspirecafe.Shared/HttpClients/IProductHttpClient.cs
@@ -1,12 +1,13 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AspireCafe.Shared.Models.Service.Product;
+using AspireCafe.Shared.Models.View.Product;
+using AspireCafe.Shared.Results;
+using Refit;
 
 namespace AspireCafe.Shared.HttpClients
 {
     public interface IProductHttpClient
     {
+        [Post("catalog/fetchmetadata")]
+        Task<Result<ProductMetaDataServiceModel>> FetchMetadata(ProductMetaDataViewModel products);
     }
 }
