+++ src/Aspirecafe/Aspirecafe.Shared/Models/Service/Counter/OrderHeaderServiceModel.cs
@@ -1,0 +1,9 @@
+﻿namespace AspireCafe.Shared.Models.Service.Counter
+{
+    public class OrderHeaderServiceModel
+    {
+        public string OrderType { get; set; }
+        public int? TableNumber { get; set; }
+        public string? CustomerName { get; set; }
+    }
+}
