+++ src/Aspirecafe/Aspirecafe.Shared/Models/Service/Counter/OrderLineItemServiceModel.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Service
+namespace AspireCafe.Shared.Models.Service.Counter
 {
     public class OrderLineItemServiceModel
     {
