+++ src/Aspirecafe/Aspirecafe.Shared/Models/Service/Counter/OrderServiceModel.cs
@@ -1,12 +1,4 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Domain;
-using AspireCafe.Shared.Models.Service;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace AspireCafe.CounterApiDomainLayer.Managers.Models.Service
+﻿namespace AspireCafe.Shared.Models.Service.Counter
 {
     public class OrderServiceModel:ServiceBaseModel
     {
