+++ src/Aspirecafe/Aspirecafe.Shared/Models/Service/Product/CatalogServiceModel.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace AspireCafe.ProductApiDomainLayer.Managers.Models.Service
+namespace AspireCafe.Shared.Models.Service.Product
 {
     public class CatalogServiceModel: ServiceBaseModel
     {
