+++ src/Aspirecafe/Aspirecafe.Shared/Models/Service/Product/ProductMetaDataServiceModel.cs
@@ -1,0 +1,9 @@
+﻿using AspireCafe.Shared.Enums;
+
+namespace AspireCafe.Shared.Models.Service.Product
+{
+    public class ProductMetaDataServiceModel: ServiceBaseModel
+    {
+        public Dictionary<Guid,RouteType> Metadata { get; set; } = new Dictionary<Guid, RouteType>();
+    }
+}
