+++ src/Aspirecafe/Aspirecafe.Shared/Models/View/Counter/OrderPaymentViewModel.cs
@@ -1,12 +1,7 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Enums;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AspireCafe.Shared.Enums;
 
-namespace AspireCafe.CounterApiDomainLayer.Managers.Models.View
-{
+namespace AspireCafe.Shared.Models.View.Counter
+{ 
     public class OrderPaymentViewModel
     {
         public Guid OrderId { get; set; }
