+++ src/Aspirecafe/Aspirecafe.Shared/Models/View/Counter/OrderViewModel.cs
@@ -1,12 +1,6 @@
-﻿using AspireCafe.CounterApiDomainLayer.Managers.Models.Enums;
-using AspireCafe.Shared.Models.View;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using AspireCafe.Shared.Enums;
 
-namespace AspireCafe.CounterApiDomainLayer.Managers.Models.View
+namespace AspireCafe.Shared.Models.View.Counter
 {
     public class OrderViewModel:ViewModelBase
     {
